==> 01-data-summary.Rmd <==
---
title: 'CJ meta-analysis: data summary'
author: "George Kinnear"
date: "2024-02-12"
output: github_document
always_allow_html: true
---

```{r setup, include=FALSE}
library(tidyverse)
knitr::opts_chunk$set(echo = FALSE)

# for plotting
theme_set(theme_minimal())
#library(patchwork)

library(knitr)
library(kableExtra)
basic_kable = function(df) {
  df %>% 
    kable() %>%
    kable_styling(bootstrap_options = "striped", full_width = F)
}
```

We gathered comparative judgement data from a range of research studies.

```{r}
# Read in the data about sessions
sessions <- yaml::read_yaml("data/00-projects-with-data.yml")

all_sessions <- tibble(project = sessions) %>% 
  unnest_wider(project) %>% 
  mutate(open_data = case_when(
    open_data == FALSE ~ "FALSE",
    .default = as.character(open_data)
  )) %>%
  unnest_longer(judging_sessions) %>%
  unnest_wider(judging_sessions) %>%
  # replace columns containing length-1 lists with just the contents
  mutate(across(c("study", "N_A", "N_R", "N_J", "adaptivity"), as.character)) %>% 
  # tidy up the missing entries - replace NULL with NA
  mutate(across(c("study", "N_A", "N_R", "N_J", "adaptivity"), ~ na_if(.x, "NULL")))
```

* This document summarises the sample in a readable way.
* It draws on raw data in the `data/00-projects-with-data.yml` file.
* This script processes the raw data and saves it in CSV format in `data/00-judging_sessions_summary.csv`.

# Publications/projects

Judgement data was gathered from the publications/projects summarised in the table below.

```{r}
all_sessions %>% 
  select(category, project_id, citation, doi, open_data) %>% 
  mutate(doi = if_else(str_detect(doi, "doi.org"), doi, paste0("https://doi.org/", doi))) %>% 
  mutate(doi_link = text_spec(doi, link = doi)) %>% 
  select(-doi) %>% 
  distinct() %>%
  arrange(category, project_id) %>% 
  kable("html", escape = FALSE) %>%
  kable_styling() 
```

## Open data

`open_data` records whether the judgement data is already publicly available. It is fairly uncommon in our sample:

```{r}
all_sessions %>% 
  select(project_id, open_data) %>% 
  distinct() %>% 
  mutate(open_data = ifelse(open_data != "FALSE", "Yes", "No")) %>% 
  janitor::tabyl(open_data) %>% 
  janitor::adorn_pct_formatting() %>% 
  basic_kable()
```

## Categories

`category` records whether the study was about "education" or "other":

```{r}
all_sessions %>% 
  select(project_id, category) %>% 
  distinct() %>% 
  mutate(category = ifelse(category != "education", "Other", "Education")) %>% 
  janitor::tabyl(category) %>% 
  janitor::adorn_pct_formatting() %>% 
  basic_kable()
```

## Sources

`source` records whether the study originated in the literature search or from subsequent "snowballing":

```{r}
all_sessions %>% 
  select(project_id, source) %>% 
  distinct() %>% 
  mutate(source = if_else(source == "snowballing", "Snowballing", "Literature search", missing = "Literature search")) %>% 
  janitor::tabyl(source) %>% 
  janitor::adorn_pct_formatting() %>% 
  basic_kable()
```

# CJ sessions

Many of the publications report on multiple CJ sessions:

```{r}
all_sessions %>% 
  select(project_id, study) %>%
  group_by(project_id) %>% 
  summarise(
    studies = paste0(study, collapse = "<br>")
  ) %>% 
  mutate(studies = ifelse(studies == "NA", "", studies)) %>%  
  kable("html", escape = FALSE) %>%
  kable_styling()
```

## Reading the data

All the judgement data is stored in the `data` folder, with filenames corresponding to the `project_id` (and for papers with multiple CJ sessions, the `study` within that project, e.g. `Bisson2016_stats.csv`).

```{r message=FALSE, warning=FALSE}
source("00-load-all-judgement-data.R")

judgement_data_tidy %>% 
  select(-csv_contents, -path) %>% 
  write_csv("data/00-judging_sessions_summary.csv")
```

### Data matches with the paper

These ones all have $N_R$ and $N_A$ numbers that match our records of what is reported in the paper:

```{r}
judgement_data_tidy %>% 
  select(judging_session, contains("N_", ignore.case = FALSE)) %>% 
  filter(!is.na(observed_N_R), !is.na(observed_N_A)) %>% 
  filter(observed_N_R == N_R & observed_N_A == N_A) %>% 
  basic_kable()
```

### Data inconsistent with reported values

The following sessions have $N_A$ or $N_R$ values that differ from those expected:

```{r}
judgement_data_tidy %>% 
  select(judging_session, contains("N_", ignore.case = FALSE)) %>% 
  filter(!is.na(observed_N_R), !is.na(observed_N_A)) %>% 
  filter(observed_N_R != N_R | observed_N_A != N_A) %>% 
  basic_kable()
```

### Data is missing

The following sessions have missing data:

```{r}
judgement_data_tidy %>% 
  filter(is.na(observed_N_C)) %>% 
  select(judging_session, citation, items, judge_expertise, note) %>% 
  basic_kable()
```

### Other issues

Various other unresolved issues are noted with a "TODO" in the project's `note` field:

```{r}
judgement_data_tidy %>% 
  filter(str_detect(note, "TODO")) %>% 
  select(judging_session, note) %>%
  kable("html", escape = FALSE) %>%
  kable_styling()
```


# Other things to do

## Adaptivity

We wanted to add a note about whether the CJ method used adaptivity.

```{r}
judgement_data_tidy %>% 
  select(judging_session, adaptivity) %>%
  janitor::tabyl(adaptivity) %>% 
  kable("html", escape = FALSE) %>%
  kable_styling()
```

## Judge expertise

```{r}
judgement_data_tidy %>% 
  select(judging_session, judge_expertise) %>% 
  janitor::tabyl(judge_expertise) %>% 
  kable("html", escape = FALSE) %>%
  kable_styling()
```



## Projects without data

Do we want to fill out details for these studies?

```{r}
# import projects-without-data
nodata <- yaml::read_yaml("data/00-projects-without-data.yml")

nodata_sessions <- tibble(project = nodata) %>% 
  unnest_wider(project) %>% 
  mutate(open_data = paste0(open_data)) %>% 
  unnest_longer(judging_sessions) %>%
  unnest_wider(judging_sessions, names_repair = "unique") %>%
  # replace columns containing length-1 lists with just the contents
  mutate(across(c("study", "N_A", "N_R", "N_J", "adaptivity"), as.character)) %>% 
  # tidy up the missing entries - replace NULL with NA
  mutate(across(c("study", "N_A", "N_R", "N_J", "adaptivity"), ~ na_if(.x, "NULL")))
```

At least, we might want to trawl through looking for open data? We included Jones2020 on this basis, see <https://github.com/georgekinnear/cj-meta-analysis/issues/2>

```{r}
# count entries with TODO in the open data field
nodata_sessions %>% janitor::tabyl(open_data)
```

# Summary table

```{r}
summary_table <- all_sessions %>%
  transmute(
    source = replace_na(source, "Search"),
    source = str_to_title(source),
    judge_expertise = str_to_title(judge_expertise),
    adaptivity = case_when(
      adaptivity == "FALSE" ~ "No",
      adaptivity == "TRUE" ~ "Yes",
      .default = "Unknown"
    ),
    category = case_when(category == "education" ~ "Education",
                         .default = "Other"),
    open_data = case_when(open_data == "FALSE" ~ "No",
                          .default = "Yes"),
  ) %>% 
  gtsummary::tbl_summary(
    by = source,
    label = list(
      adaptivity ~ "Adaptive CJ",
      judge_expertise ~ "Assessor expertise",
      open_data ~ "Open data"
    )
  ) %>% 
  gtsummary::add_overall() %>% 
  gtsummary::modify_footnote(update = everything() ~ NA)

summary_table
```

<details><summary>LaTeX table</summary>

```{r}
summary_table %>% gtsummary::as_gt() %>% gt::as_latex() %>% cat()
```
</details>
